Extract element from entrypoint
